+++ src/core/models/fluorescence/fl_2_1_1_1_Model.h
@@ -62,7 +62,8 @@ public:
     }
 
     void virtual DeclareOptions() override;
-    virtual int LocalParameterSize() const override { return 4; }
+    virtual int LocalParameterSize(int series = 0) const override { Q_UNUSED(series)
+        return 4; }
     void virtual EvaluateOptions() override;
 
 protected: