+++ src/core/models/itc/blank.h
@@ -54,7 +54,8 @@ public:
             return QString();
     }
 
-    virtual int LocalParameterSize() const override { return 2; }
+    virtual int LocalParameterSize(int i = 0) const override { Q_UNUSED(i)
+        return 2; }
     virtual inline int InputParameterSize() const override { return 1; }
 
     virtual inline int Color(int i) const override