+++ src/core/models/itc/itc_1_1_Model.h
@@ -78,7 +78,8 @@ public:
             return QString();
     }
 
-    virtual int LocalParameterSize() const override { return 4; }
+    virtual int LocalParameterSize(int i = 0) const override { Q_UNUSED(i)
+        return 4; }
     virtual inline int InputParameterSize() const override { return 1; }
 
     virtual inline int Color(int i) const override