+++ src/core/models/itc/itc_2_1_Model.h
@@ -89,7 +89,8 @@ public:
             return QString();
     }
 
-    virtual int LocalParameterSize() const override { return 5; }
+    virtual int LocalParameterSize(int i = 0) const override { Q_UNUSED(i)
+        return 5; }
     virtual inline int InputParameterSize() const override { return 1; }
 
     virtual inline int Color(int i) const override