+++ Changelog.md
@@ -1,5 +1,7 @@
 # SupraFit 2.0 pre-Alpha
 
+- Add global search for local parameter in series, partly for meta models as well
+- Some rework of the global search results table
 - Import of thermogram from plain txt files possible, peak time and peak start have to be given for peak picking
 - Improved thermodynamic output -> calculation of free enthalpy and entropy (if heat available) including confidence
 - ToolSet::Confidence functions gives identical results compared to the octave quantile function